+++ src/main/java/com/journal/florist/backend/feature/product/dto/product/UpdateProductRequest.java
@@ -20,7 +20,7 @@ public class UpdateProductRequest implements Serializable {
     private Integer stock;
     private BigDecimal price;
     private String categoryId;
-    private Integer weight;
+    private Long weight;
     private String material;
     private BigDecimal materialPrice;
 }
